Moved utility functions to different source file in jablkomods
